+++ config/environments/production.rb
@@ -83,17 +83,15 @@ Rails.application.configure do
   # Do not dump schema after migrations.
   config.active_record.dump_schema_after_migration = false
 
-  config.action_mailer.default_url_options = {:host => 'http://store-finding-system.herokuapp.com' }
+  config.action_mailer.default_url_options = {:host => 'http://dinobo.com' }
 
-  config.action_mailer.delivery_method = :smtp
   config.action_mailer.smtp_settings = {
-    enable_starttls_auto: true,
-    address: "smtp.gmail.com",
-    port: 587,
-    domain: "imap.gmail.com",
-    authentication: :login,
-    user_name: "rubyrails9@gmail.com",
-    password: "utmeter!@#02"
+    :address   => "smtp.mandrillapp.com",
+    :port      => 587,
+    :user_name => ENV.fetch('MANDRILL_USERNAME'),
+    :password  => ENV.fetch('MANDRILL_API_KEY'),
+    :domain    => 'dinobo.com'
   }
+  config.action_mailer.delivery_method = :smtp
 
 end